+++ Assets/HTC.UnityPlugin/VRModule/Modules/UnityXRModuleUtil.cs
@@ -161,7 +161,7 @@ namespace HTC.UnityPlugin.VRModuleManagement
         public static bool HasActiveLoader(VRModuleKnownXRLoader knownLoader)
         {
             string activeLoaderName;
-            return TryGetActiveLoaderName(out activeLoaderName) && knownLoader == ToKnownXRLoader(activeLoaderName);
+            return TryGetActiveLoaderName(out activeLoaderName) && (knownLoader == ToKnownXRLoader(activeLoaderName));
         }
 
         public static VRModuleKnownXRLoader GetKnownActiveLoader()
@@ -178,14 +178,7 @@ namespace HTC.UnityPlugin.VRModuleManagement
             var loaderName = loader.name;
             if (string.IsNullOrEmpty(loaderName)) { return VRModuleKnownXRLoader.Unknown; }
 
-            foreach (var profile in loaderProfiles)
-            {
-                if (string.IsNullOrEmpty(profile.fixedName) || profile.fixedName != loaderName) { continue; }
-                if (profile.matchNameRgx == null || !profile.matchNameRgx.IsMatch(loaderName)) { continue; }
-                return profile.loader;
-            }
-
-            return VRModuleKnownXRLoader.Unknown;
+            return ToKnownXRLoader(loaderName);
         }
 
         public static VRModuleKnownXRLoader ToKnownXRLoader(string loaderName)
@@ -246,7 +239,11 @@ namespace HTC.UnityPlugin.VRModuleManagement
             var activeSubsys = ListPool<XRInputSubsystem>.Get();
             try
             {
+#if UNITY_6000_0_OR_NEWER
+                SubsystemManager.GetSubsystems(activeSubsys);
+#else
                 SubsystemManager.GetInstances(activeSubsys);
+#endif
                 if (activeSubsys.Count == 0)
                 {
                     Debug.LogWarning("No XRInputSubsystem detected.");
@@ -257,7 +254,11 @@ namespace HTC.UnityPlugin.VRModuleManagement
                     {
                         if (!subsys.running) { continue; }
 
+#if UNITY_6000_0_OR_NEWER
+                        var subsysName = subsys.subsystemDescriptor.id;
+#else
                         var subsysName = subsys.SubsystemDescriptor.id;
+#endif
                         if (string.IsNullOrEmpty(subsysName)) { continue; }
 
                         foreach (var profile in inputSubsystemProfiles)