+++ Assets/HTC.UnityPlugin/ViveInputUtility/Examples/8.NearFieldHandInteraction/Scripts/PushButton.cs
@@ -72,7 +72,11 @@ namespace HTC.UnityPlugin.Vive
             // Lock velocity
             if (m_rigidbody)
             {
+#if UNITY_6000_0_OR_NEWER
+                m_rigidbody.linearVelocity = Vector3.zero;
+#else
                 m_rigidbody.velocity = Vector3.zero;
+#endif
             }
         }