+++ Assets/HTC.UnityPlugin/ViveInputUtility/Scripts/Editor/VIUSettingsEditor.cs
@@ -427,9 +427,8 @@ namespace HTC.UnityPlugin.Vive
                         case StatusCode.Failure:
                             if (!s_wasRemoved)
                             {
-                                var request = m_removeRequest;
-                                m_removeRequest = null;
                                 Debug.LogError("Something wrong when removing package from list. error:" + m_removeRequest.Error.errorCode + "(" + m_removeRequest.Error.message + ")");
+                                m_removeRequest = null;
                             }
                             break;
                         case StatusCode.Success:
@@ -756,7 +755,9 @@ namespace HTC.UnityPlugin.Vive
                 //    EditorGUILayout.HelpBox("AndroidNdkRoot is empty. Setup at Edit -> Preferences... -> External Tools -> Android SDK", MessageType.Warning);
                 //}
 
-#if UNITY_5_6_OR_NEWER && !UNITY_5_6_0 && !UNITY_5_6_1 && !UNITY_5_6_2
+#if UNITY_6000_0_OR_NEWER
+                if (PlayerSettings.GetApplicationIdentifier(UnityEditor.Build.NamedBuildTarget.Android).Equals("com.Company.ProductName"))
+#elif UNITY_5_6_OR_NEWER && !UNITY_5_6_0 && !UNITY_5_6_1 && !UNITY_5_6_2
                 if (PlayerSettings.GetApplicationIdentifier(BuildTargetGroup.Android).Equals("com.Company.ProductName"))
 #else
                 if (PlayerSettings.bundleIdentifier.Equals("com.Company.ProductName"))